+++ packages/plugin-commons/web/src/index.ts
@@ -31,9 +31,6 @@ export { default as createBaseInsertPluginButton } from './Base/createBaseInsert
 export { default as baseToolbarButton } from './Base/baseToolbarButton';
 
 // Components
-export { default as RadioGroupHorizontal } from './Components/RadioGroupHorizontal';
-export { default as RadioGroupVertical } from './Components/RadioGroupVertical';
-export { default as SettingsPanelFooter } from './Components/SettingsPanelFooter';
 export { default as ColorPicker } from './Components/ColorPicker/ColorPicker';
 
 export { getSelectionStyles } from './Utils/inlineStyleUtils';