+++ packages/ui-components/web/src/index.ts
@@ -2,19 +2,18 @@
 
 // } from './consts';
 
-// // Components
+// Components
 export { default as Button } from './Components/Button';
 export { default as FileInput } from './Components/FileInput';
-// export { default as FileInput } from './Components/FileInput';
 export { default as Image } from './Components/Image';
 export { default as InputWithLabel } from './Components/InputWithLabel';
 export { default as LabeledToggle } from './Components/LabeledToggle';
 export { default as Panel } from './Components/Panel';
 export { default as SliderPanel } from './Components/SliderPanel';
-// export { default as RadioGroupHorizontal } from './Components/RadioGroupHorizontal';
-// export { default as RadioGroupVertical } from './Components/RadioGroupVertical';
+export { default as RadioGroupVertical } from './Components/RadioGroupVertical';
+export { default as RadioGroupHorizontal } from './Components/RadioGroupHorizontal';
 export { default as SelectionList } from './Components/SelectionList';
-// export { default as SettingsPanelFooter } from './Components/SettingsPanelFooter';
+export { default as SettingsPanelFooter } from './Components/SettingsPanelFooter';
 export { default as SettingsSection } from './Components/SettingsSection';
 export { default as SliderWithInput } from './Components/SliderWithInput';
 export { default as Slider } from './Components/Slider';